# openapi.yml
openapi: 3.0.0
info:
  title: Campusoul API
  description: API for a Campusoul application including user management, interest management, image handling, matches, and messaging.
  version: "1.0.0"
servers:
  - url: http://localhost:3000/
...
paths:
...
  /images:
    post:
      tags: [Images]
      summary: Upload an image
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Image uploaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: No file uploaded
        "500":
          description: Server error or error uploading to S3
  /images/{imageId}:
    get:
      tags: [Images]
      summary: Retrieve an image by ID
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ImageId"
      responses:
        "200":
          description: Image data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Image"
        "404":
          description: Image not found
        "500":
          description: Server error

    delete:
      tags: [Images]
      summary: Delete an image by ID
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ImageId"
      responses:
        "200":
          description: Image deleted successfully
        "404":
          description: Image not found
        "500":
          description: Server error or error deleting from S3
...
  /messages/send:
    post:
      tags: [Messages]
      summary: Send a message
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                matchId:
                  type: string
                content:
                  type: string
      responses:
        "201":
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "500":
          description: Server error
  /messages/{matchId}:
    get:
      tags: [Messages]
      summary: Retrieve messages for a match
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/MatchId"
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
        "500":
          description: Server error
components:
  schemas:
    User:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        bio:
          type: string
        birthdate:
          type: string
          format: date
        interests:
          type: array
          items:
            $ref: "#/components/schemas/Interest"
        isAdmin:
          type: boolean
        location:
          type: object
          properties:
            type:
              type: string
              enum: ["Point"]
            coordinates:
              type: array
              items:
                type: number
        images:
          type: array
          items:
            $ref: "#/components/schemas/Image"
          createdAt:
            type: string
            format: date-time
        interest:
          type: array
          items:
            $ref: "#/components/schemas/Interest"
...
    Match:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        matchedAt:
          type: string
          format: date-time
        isMatchActive:
          type: boolean
    Message:
      type: object
      properties:
        match:
          $ref: "#/components/schemas/Match"
        sender:
          $ref: "#/components/schemas/User"
        receiver:
          $ref: "#/components/schemas/User"
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        read:
          type: boolean
    Image:
      type: object
      properties:
        url:
          type: string
        createdAt:
          type: string
          format: date-time
    Interest:
      type: object
      properties:
        name:
          type: string
        createdAt:
          type: string
          format: date-time
  parameters:
...
    MatchId:
      name: matchId
      in: path
      required: true
      schema:
        type: string
    ImageId:
      name: imageId
      in: path
      required: true
      schema:
        type: string
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT